pyproject.toml
[project]
name = "bar-benchmarks"
version = "0.1.0"
description = "GCP Batch harness for running Beyond All Reason performance scenarios across identical ephemeral VMs."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "typer>=0.12",
    "google-cloud-batch>=0.17",
    "google-cloud-storage>=2.16",
]

[project.scripts]
bar-bench = "bar_benchmarks.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bar_benchmarks"]

[dependency-groups]
dev = [
    "pytest>=8",
    "ruff>=0.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
